Run permutation test on PCA analysis.

Permute data fed to PCA a given number of times, collecting the number of significant pairwise correlations in the permuted data and the variances explained for a given number of PCs.

Usage

permutation_test(
  pca_data,
  pc_n = 5,
  n = 100,
  scale = TRUE,
  cor.method = "pearson"
)

Arguments

pca_data

data fed to the prcomp function. Remove non-continuous variables.

pc_n

the number of PCs to collect variance explained from.

n

the number of times to permute that data. Warning: high values will take a long time to compute.

scale

whether the PCA variables should be scaled (default = TRUE).

cor.method

method to use for correlations (default = "pearson"). Alternative is "spearman".

Value

object of class permutation_test

$permuted_variances n x pc_no matrix of variances explained by first pc_no PCs in n permutations of original data.

$permuted_correlations list of length n of significant pairwise correlations in n permutations of the data (<= 0.05).

$actual_variances pc_n x 2 tibble of variances explained by first pc_n PCs with original data.

$actual_correlations the number of significant pairwise correlations (<= 0.05) in the original data.

Details

Use correlation_test() for pairwise correlations and pca_test() for variance explained and loadings.

Examples

permutation_test(
  onze_intercepts |> dplyr::select(-speaker),
  pc_n = 5,
  n = 10,
  scale = TRUE,
  cor.method = 'pearson'
 )
#> $permuted_variances
#>           PC1        PC2        PC3        PC4        PC5
#> 1  0.10035139 0.08710470 0.08323354 0.07762683 0.06760223
#> 2  0.09003776 0.08305271 0.07842767 0.07104812 0.06974112
#> 3  0.09991849 0.09113667 0.07798557 0.07564895 0.06550545
#> 4  0.10473515 0.08704997 0.08123940 0.07906010 0.07433931
#> 5  0.09804940 0.08999558 0.08015930 0.07472181 0.06995225
#> 6  0.09346045 0.08768051 0.08209320 0.07410834 0.06661119
#> 7  0.09484123 0.08416786 0.07542503 0.07086044 0.06603212
#> 8  0.09979899 0.09583668 0.07978208 0.07523975 0.07417293
#> 9  0.09041724 0.08375506 0.07829196 0.07310492 0.06588910
#> 10 0.10192328 0.08706154 0.07614326 0.07370574 0.06930026
#> 
#> $permuted_correlations
#>  [1] 11 10 14 19 16 10  4 16  7 14
#> 
#> $actual_variances
#> # A tibble: 5 × 2
#>   PC    variance_explained
#>   <chr>              <dbl>
#> 1 PC1               0.183 
#> 2 PC2               0.143 
#> 3 PC3               0.0989
#> 4 PC4               0.0835
#> 5 PC5               0.0774
#> 
#> $actual_correlations
#> [1] 69
#> 
#> attr(,"class")
#> [1] "permutation_results"